+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Changed
+* Improved accuracy of material assignments.
+* Changed material import to automatically arrange shader nodes.
+* Adjusted wimdo import and export to preserve the 4th component from the "VertexNormal" attribute used for normal map intensity for some models.
+
 ## 0.19.0 - 2025-03-28
 ### Added
 * Added support for importing Xenoblade Chronicles X Definitive Edition wimdo models.